refactor(cli): extract AgentRuntime and Banner into dedicated modules
