+++ pyproject.toml
@@ -14,6 +14,7 @@ dependencies = [
 [dependency-groups]
 dev = [
     "ipython>=8.38.0",
+    "langsmith>=0.7.22",
     "pytest>=9.0.2",
     "pytest-mock>=3.15.1",
 ]
